+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added menu option to show Readme
 - Update wellen to 0.25.6
 - Fix glitch with Alt + Down next to collapsed groups
+- Delta Time in the status bar now displays units in Hz
 
 ## 1.5.4 - 6/4/2026 - Latest Release